requirements.txt
asgiref==3.2.10
autopep8==1.5.7
Django==3.1.1
django-bootstrap-form==3.4
django-ckeditor==6.0.0
django-filter==2.3.0
django-js-asset==1.2.2
flake8==3.9.2
mccabe==0.6.1
Pillow==8.1.2
pycodestyle==2.7.0
pyflakes==2.3.1
python-decouple==3.4
pytz==2020.1
sqlparse==0.3.1
toml==0.10.2
